[Ingest] Move some server lifecycle types out of common. (#64252)
The AgentService, ESIndexPatternService, and IngestManager{Setup,Start}{Deps,Contract} values were in common, but they are server only.
