+++ x-pack/plugins/endpoint/server/mocks.ts
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { AgentService, IngestManagerStartupContract } from '../../ingest_manager/common/types';
+import { AgentService, IngestManagerStartContract } from '../../ingest_manager/server';
 
 /**
  * Creates a mock IndexPatternRetriever for use in tests.
@@ -46,9 +46,9 @@ export const createMockAgentService = (): jest.Mocked<AgentService> => {
  * @param indexPattern a string index pattern to return when called by a test
  * @returns the same value as `indexPattern` parameter
  */
-export const createMockIngestManagerStartupContract = (
+export const createMockIngestManagerStartContract = (
   indexPattern: string
-): IngestManagerStartupContract => {
+): IngestManagerStartContract => {
   return {
     esIndexPatternService: {
       getESIndexPattern: jest.fn().mockResolvedValue(indexPattern),
