+++ x-pack/plugins/ingest_manager/common/types/index.ts
@@ -3,42 +3,8 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import { SavedObjectsClientContract } from 'kibana/server';
-import { AgentStatus } from './models';
-
 export * from './models';
 export * from './rest_spec';
-
-/**
- * Service to return the index pattern of EPM packages
- */
-export interface ESIndexPatternService {
-  getESIndexPattern(
-    savedObjectsClient: SavedObjectsClientContract,
-    pkgName: string,
-    datasetPath: string
-  ): Promise<string | undefined>;
-}
-
-/**
- * Describes public IngestManager plugin contract returned at the `startup` stage.
- */
-export interface IngestManagerStartupContract {
-  esIndexPatternService: ESIndexPatternService;
-  agentService: AgentService;
-}
-
-/**
- * A service that provides exported functions that return information about an Agent
- */
-export interface AgentService {
-  /**
-   * Return the status by the Agent's id
-   * @param soClient
-   * @param agentId
-   */
-  getAgentStatusById(soClient: SavedObjectsClientContract, agentId: string): Promise<AgentStatus>;
-}
 
 export interface IngestManagerConfigType {
   enabled: boolean;
