+++ x-pack/plugins/ingest_manager/server/index.ts
@@ -6,6 +6,12 @@
 import { schema, TypeOf } from '@kbn/config-schema';
 import { PluginInitializerContext } from 'src/core/server';
 import { IngestManagerPlugin } from './plugin';
+export { AgentService, ESIndexPatternService } from './services';
+export {
+  IngestManagerSetupContract,
+  IngestManagerSetupDeps,
+  IngestManagerStartContract,
+} from './plugin';
 
 export const config = {
   exposeToBrowser: {
